requirements.txt
datasets==2.14.5
Kivy==2.2.1
peft==0.5.0
PyYAML==6.0.1
PyYAML==6.0.1
torch==2.1.0
transformers==4.34.1
